requirements.txt
boto3
python-dotenv
PyJWT
bcrypt
redis
uuid